Slight refactoring, fixed bug that raises errors when len(x) != len(y) (pixels)
