+++ tests/test_gpu.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pytest
 
-from deinterlacing.processing import DeinterlaceParameters, deinterlace
+from deinterlacing import DeinterlaceParameters
+from deinterlacing.processing import deinterlace
 from tests.conftest import acquire_gpu, release_gpu
